requirements.txt
pillow==10.2.0
tqdm==4.66.1
torch==2.2.0
ultralytics==8.1.0
numpy>=1.24.0,<2.0.0
opencv-python==4.9.0.80
pandas>=2.0.0
pyyaml>=6.0.1
shapely>=2.0.0